refactor(acms): Extract ACMS API helpers to background service
